+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,124 @@
 
 ### Patch Changes
 
+- Updated dependencies [53f6731]
+  - ai@6.0.67
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.66
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.65
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [ce9daa3]
+  - ai@6.0.64
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [be95579]
+  - ai@6.0.63
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [2810850]
+  - ai@6.0.62
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.61
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [5fc42fa]
+  - ai@6.0.60
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.59
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.58
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [65865d8]
+  - ai@6.0.57
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.56
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [43a74df]
+  - ai@6.0.55
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [2f8ac87]
+  - ai@6.0.54
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [7ee3f10]
+  - ai@6.0.53
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.52
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [ea0feb5]
+  - ai@6.0.51
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.50
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [ded661b]
   - ai@6.0.49